整合部分协议

+++ aiiopenapi/src/main/java/com/aiitec/openapi/model/FileListResponseQuery.java
@@ -1,18 +1,15 @@
 package com.aiitec.openapi.model;
 
-import android.os.Parcel;
-
-import java.util.ArrayList;
-import java.util.List;
-
 import com.aiitec.openapi.json.annotation.JSONField;
+
+import java.util.List;
 
 
 public class FileListResponseQuery extends ListResponseQuery {
 
 	@JSONField(entityName="file")
 	private List<File> files;
-	private List<Integer> ids;
+	private List<Long> ids;
 
     public List<File> getFiles() {
         return files;
@@ -22,11 +19,11 @@ public class FileListResponseQuery extends ListResponseQuery {
         this.files = files;
     }
 
-    public List<Integer> getIds() {
+    public List<Long> getIds() {
         return ids;
     }
 
-    public void setIds(List<Integer> ids) {
+    public void setIds(List<Long> ids) {
         this.ids = ids;
     }
